% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afnor_azerty}
\alias{afnor_azerty}
\title{Afnor AZERTY keyboard}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 119 rows and 19 columns.
}
\usage{
afnor_azerty
}
\description{
A dataset containing rows and keys of a full ISO Afnor AZERTY layout.
}
\keyword{datasets}
