% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch_qwertz}
\alias{ch_qwertz}
\title{Swiss French keyboard}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 119 rows and 19 columns.
}
\usage{
ch_qwertz
}
\description{
A dataset containing rows and keys of a full ISO Swiss French layout, as commonly used in Luxembourg
}
\keyword{datasets}
