% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{compile_rules}
\alias{compile_rules}
\title{Compile rules for optimizer}
\usage{
compile_rules(rules, layout, keyboard)
}
\arguments{
\item{rules}{List of rule objects created by rule builder functions.}

\item{layout}{Character vector of initial layout (26 keys).}

\item{keyboard}{Keyboard data frame with position info.}
}
\value{
A list with compiled rule data for C++.
}
\description{
Internal function that compiles a list of rule objects into the format
expected by the C++ optimizer.
}
\keyword{internal}
