% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{english}
\alias{english}
\title{Sample of english text}
\format{
An object of class \code{character} of length 51.
}
\usage{
english
}
\description{
A sample of english text
}
\keyword{datasets}
