% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{fix_keys}
\alias{fix_keys}
\title{Fix keys in their initial positions}
\usage{
fix_keys(keys)
}
\arguments{
\item{keys}{Character vector of keys to fix in place.}
}
\value{
A rule object of class "layout_rule".
}
\description{
Creates a hard constraint that prevents specified keys from moving
during optimization. Useful for keeping shortcut keys (Z, X, C, V)
in familiar positions.
}
\examples{
\dontrun{
# Keep Ctrl shortcut keys in place
rule <- fix_keys(c("z", "x", "c", "v"))
}
}
