% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapize.R
\name{heatmapize}
\alias{heatmapize}
\title{Colours the keys of the keyboard according the relative frequency of characters appearing in a text.}
\usage{
heatmapize(keyboard, letter_freq_df, low = "light green", high = "red")
}
\arguments{
\item{keyboard}{A dataframe holding a keyboard specification}

\item{letter_freq_df}{A dataframe with relative frequencies of characters. Output of \code{letter_freq()}}

\item{low}{A colour specification for low frequencies. Can be a hex colour code or one of the inbuilt colours. Defaults to "light green".}

\item{high}{A colour specification for high frequencies. Can be a hex colour code or one of the inbuilt colours. Defaults to "red".}
}
\value{
The original keyboard dataframe but with an adjusted \code{fill} column, which colours each character according to its relative frequency.
}
\description{
Colours the keys of the keyboard according the relative frequency of characters appearing in a text.
}
\examples{
\dontrun{
# Create some random string
text <- sample(replicate(10, letters), replace = TRUE)
letter_freq_df <- letter_freq(text)
heatmap_azerty <- heatmapize(afnor_azerty, letter_freq_df)
}
}
