% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter_freq.R
\name{letter_freq}
\alias{letter_freq}
\title{Counts the frequency of each letter in a string of text}
\usage{
letter_freq(text, only_alpha = TRUE)
}
\arguments{
\item{text}{A string}

\item{only_alpha}{TRUE (default) or FALSE. Should only letters be considered? If FALSE, every character is taken into account.}
}
\value{
A dataframe with 4 columns. \code{characters} contains the characters; \code{total} the total number of times the characters appears in the text, \code{scaled} is the min-max transform of \code{total} and \code{frequencies} are the relative frequencies of appearance of each letter.
}
\description{
Counts the frequency of each letter in a string of text
}
\examples{
\dontrun{
# Create some random string
data("french")
letter_freq(french)
}
}
