% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_layout_ini.R
\name{list_layouts}
\alias{list_layouts}
\title{List available layouts in a directory}
\usage{
list_layouts(path = NULL)
}
\arguments{
\item{path}{Directory to search for layout.ini files.
Default is "layouts" subdirectory of the current package.}
}
\value{
A character vector of paths to layout.ini files.
}
\description{
Finds all layout.ini files in a given directory and its subdirectories.
}
\examples{
\dontrun{
# List layouts in the package
layouts <- list_layouts()
print(layouts)
}
}
