% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luxembourguish}
\alias{luxembourguish}
\title{Sample of luxembourguish text}
\format{
An object of class \code{character} of length 188.
}
\usage{
luxembourguish
}
\description{
A sample of luxembourguish text
}
\keyword{datasets}
