% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{min_max}
\alias{min_max}
\title{Standardizes an atomic vector of numbers using the min-max method}
\usage{
min_max(x)
}
\arguments{
\item{x}{An atomic vector of numbers}
}
\description{
Standardizes an atomic vector of numbers using the min-max method
}
\examples{
\dontrun{
min_max(seq(1, 10))
}
}
