% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{print.layout_rule}
\alias{print.layout_rule}
\title{Print method for layout rules}
\usage{
\method{print}{layout_rule}(x, ...)
}
\arguments{
\item{x}{A layout_rule object}

\item{...}{Ignored}
}
\description{
Print method for layout rules
}
