% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_layout_ini.R
\name{print.pkl_layout}
\alias{print.pkl_layout}
\title{Print method for pkl_layout objects}
\usage{
\method{print}{pkl_layout}(x, ...)
}
\arguments{
\item{x}{A pkl_layout object from parse_layout_ini()}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for pkl_layout objects
}
